+++ mockserver/mockserver-netty/src/test/java/org/mockserver/netty/http3/Http3GrpcStreamingIntegrationTest.java
@@ -18,6 +18,7 @@ import io.netty.handler.codec.quic.QuicChannel;
 import io.netty.handler.codec.quic.QuicSslContext;
 import io.netty.handler.codec.quic.QuicSslContextBuilder;
 import io.netty.handler.codec.quic.QuicStreamChannel;
+import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
 import org.junit.After;
 import org.junit.Assume;
 import org.junit.Before;
@@ -37,8 +38,6 @@ import org.mockserver.netty.MockServer;
 import org.mockserver.serialization.model.PausedStreamFrameDTO;
 import org.mockserver.serialization.model.StreamFrameDecisionDTO;
 
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
 import java.io.ByteArrayOutputStream;
 import java.net.InetSocketAddress;
 import java.nio.file.Path;
@@ -579,7 +578,7 @@ public class Http3GrpcStreamingIntegrationTest {
             clientGroup = new NioEventLoopGroup(1);
 
             QuicSslContext clientSslContext = QuicSslContextBuilder.forClient()
-                .trustManager(trustAllManager())
+                .trustManager(InsecureTrustManagerFactory.INSTANCE)
                 .applicationProtocols(Http3.supportedApplicationProtocols())
                 .build();
 
@@ -652,23 +651,5 @@ public class Http3GrpcStreamingIntegrationTest {
         } catch (Throwable t) {
             Assume.assumeNoException("native QUIC transport failed to load -- skipping HTTP/3 gRPC streaming test", t);
         }
-    }
-
-    @SuppressWarnings("TrustAllX509TrustManager")
-    private static TrustManager trustAllManager() {
-        return new X509TrustManager() {
-            @Override
-            public void checkClientTrusted(java.security.cert.X509Certificate[] chain, String authType) {
-            }
-
-            @Override
-            public void checkServerTrusted(java.security.cert.X509Certificate[] chain, String authType) {
-            }
-
-            @Override
-            public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-                return new java.security.cert.X509Certificate[0];
-            }
-        };
     }
 }