+++ mockserver/mockserver-netty/src/test/java/org/mockserver/netty/http3/Http3ParityIntegrationTest.java
@@ -18,6 +18,7 @@ import io.netty.handler.codec.quic.QuicChannel;
 import io.netty.handler.codec.quic.QuicSslContext;
 import io.netty.handler.codec.quic.QuicSslContextBuilder;
 import io.netty.handler.codec.quic.QuicStreamChannel;
+import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
 import org.junit.After;
 import org.junit.Assume;
 import org.junit.Before;
@@ -27,8 +28,6 @@ import org.mockserver.configuration.Configuration;
 import org.mockserver.netty.MockServer;
 import org.mockserver.socket.tls.KeyAndCertificateFactory;
 
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
 import java.net.InetSocketAddress;
 import java.nio.charset.StandardCharsets;
 import java.security.PrivateKey;
@@ -366,7 +365,7 @@ public class Http3ParityIntegrationTest {
         clientGroup = new NioEventLoopGroup(1);
 
         QuicSslContextBuilder sslBuilder = QuicSslContextBuilder.forClient()
-            .trustManager(trustAllManager())
+            .trustManager(InsecureTrustManagerFactory.INSTANCE)
             .applicationProtocols(Http3.supportedApplicationProtocols());
 
         // if client cert is provided, configure mTLS on the client side
@@ -486,23 +485,5 @@ public class Http3ParityIntegrationTest {
                 t
             );
         }
-    }
-
-    @SuppressWarnings("TrustAllX509TrustManager")
-    private static TrustManager trustAllManager() {
-        return new X509TrustManager() {
-            @Override
-            public void checkClientTrusted(java.security.cert.X509Certificate[] chain, String authType) {
-            }
-
-            @Override
-            public void checkServerTrusted(java.security.cert.X509Certificate[] chain, String authType) {
-            }
-
-            @Override
-            public java.security.cert.X509Certificate[] getAcceptedIssuers() {
-                return new java.security.cert.X509Certificate[0];
-            }
-        };
     }
 }